🧹 [Component Refactoring: DipoleWire] Extract geometry hook
